# contracts/openapi/control-plane.v1.yaml
openapi: 3.1.0
...
info:
  title: Talosmine Control Plane API
  version: 1.0.0
  description: |
    Control Plane của Talosmine — nguồn sự thật về identity, entitlement và quota.

    **Trạng thái: P1 (bootstrap).** Chỉ có health endpoint. Không có capability nghiệp vụ nào.

    Quy ước xuyên phase:
    - Số lượng (quantity/limit/counter/amount) biểu diễn bằng **decimal string**, không phải
      JSON number — PostgreSQL dùng `bigint` và JSON number mất chính xác ngoài 2^53.
    - Thời gian dùng timestamp có timezone; business time dựa DB clock.
    - Mọi lỗi trả về `ErrorEnvelope` với `code` máy đọc được. Client bắt theo `code`,
      không theo `message`.
  license:
    name: UNLICENSED
servers:
  - url: https://talosmine.localhost
    description: Local dev qua Caddy (DEC-T11)
...
paths:
...
  /v1/catalog/applications/{key}:
    get:
      tags: [catalog]
      operationId: getPublicApplication
      summary: Chi tiet mot ung dung
      description: |
        Tra theo `key` chu khong phai `id`: `key` la dinh danh on dinh ma con nguoi doc
        duoc va URL dung duoc. `id` la chi tiet noi bo.

        App chua phat hanh tra 404 GIONG HET app khong ton tai. Phan biet hai truong hop
        cho phep do xem he thong dang chuan bi nhung app nao.
      security:
        - webSession: []
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Chi tiet app.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicApplication'
        '401':
          description: Thieu phien hoac phien khong hop le.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Khong tim thay, HOAC ton tai nhung chua phat hanh.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/admin/catalog/applications:
    get:
      tags: [admin]
      operationId: adminListApplications
      summary: Danh muc ung dung (quan tri)
      description: Yeu cau `catalog:read`. Tra MOI trang thai, ke ca `draft`.
      security:
        - webSession: []
      responses:
        '200':
          description: Danh sach app, sap theo `key`.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminApplication'
        '403':
          description: Thieu permission `catalog:read`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags: [admin]
      operationId: adminCreateApplication
      summary: Tao ung dung moi
      description: |
        Yeu cau `catalog:manage`. App moi LUON o trang thai `draft` — khong tao thang
        `active`. Phat hanh la hanh dong rieng can `catalog:publish`.

        `launchUrl` va `imageUrl` di qua CHINH SACH URL truoc khi cham database: bat buoc
        https, host phai trong allowlist, chan dia chi noi bo. Xem `docs/url-policy.md`.

        URL duoc luu o DANG CHUAN HOA, khong phai chuoi goc.
      security:
        - webSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key, displayName, launchUrl, reason]
              properties:
                key:
                  type: string
                  maxLength: 64
                  pattern: '^[a-z][a-z0-9-]*$'
                  description: |
                    Machine key ON DINH va BAT BIEN. Policy request, plan grant va du lieu
                    usage tham chieu no. Khong co duong sua sau khi tao, ke ca cho quan tri.
                displayName: { type: string, maxLength: 120 }
                description: { type: [string, 'null'], maxLength: 2000 }
                imageUrl: { type: [string, 'null'], maxLength: 2048 }
                launchUrl: { type: string, maxLength: 2048 }
                reason:
                  type: string
                  minLength: 1
                  description: Bat buoc — ghi vao nhat ky kiem toan.
      responses:
        '201':
          description: Da tao, o trang thai `draft`.
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id: { type: string, format: uuid }
        '400':
          description: |
            Thieu `reason`, `key` sai dinh dang, hoac URL khong dat chinh sach (khong phai
            https / host chua allowlist / co userinfo).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Thieu permission `catalog:manage`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: '`key` da duoc dung. Key KHONG tai su dung duoc.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/admin/catalog/applications/{applicationId}:
    get:
      tags: [admin]
      operationId: adminGetApplication
      summary: Chi tiet ung dung (quan tri)
      description: Yeu cau `catalog:read`.
      security:
        - webSession: []
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        '200':
          description: Chi tiet app.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApplication'
        '403':
          description: Thieu permission `catalog:read`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Khong tim thay.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [admin]
      operationId: adminUpdateApplication
      summary: Sua metadata ung dung
      description: |
        Yeu cau `catalog:manage`.

        KHONG sua duoc `key` (bat bien) va KHONG sua duoc `status` (co duong rieng, can
        permission khac). Truong la trong body bi BO QUA im lang.

        URL moi van di qua chinh sach URL.
      security:
        - webSession: []
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                displayName: { type: string, maxLength: 120 }
                description: { type: [string, 'null'], maxLength: 2000 }
                imageUrl: { type: [string, 'null'], maxLength: 2048 }
                launchUrl: { type: string, maxLength: 2048 }
                reason: { type: string, minLength: 1 }
      responses:
        '204':
          description: Da cap nhat.
        '400':
          description: Thieu `reason` hoac URL khong dat chinh sach.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Thieu permission `catalog:manage`, hoac thieu/sai CSRF token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Khong tim thay.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/admin/catalog/applications/{applicationId}/status:
    post:
      tags: [admin]
      operationId: adminChangeApplicationStatus
      summary: Doi trang thai vong doi
      description: |
        Yeu cau `catalog:publish` — permission RIENG, khong phai `catalog:manage`.

        Vi sao tach: dua app sang `active` la mo mot `launch_url` cho nguoi dung bam vao.
        Do la hanh dong co he qua ben ngoai, khac han viec sua mo ta o trang thai nhap.
        Nguoi soan noi dung va nguoi quyet dinh phat hanh khong nhat thiet la mot.

        Chuyen tiep hop le:
            draft  -> active
            active -> inactive
            inactive -> active

        KHONG quay lai `draft`: app da tung `active` nghia la nguoi dung da thay no va co
        the da co du lieu usage. Dua ve `draft` tao ra trang thai "chua tung phat hanh"
        cho mot thu da phat hanh — dau vet lich su noi doi. Muon go thi dung `inactive`.
      security:
        - webSession: []
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status, reason]
              properties:
                status: { type: string, enum: [active, inactive] }
                reason: { type: string, minLength: 1 }
      responses:
        '204':
          description: Da doi trang thai.
        '400':
          description: Chuyen tiep khong hop le, hoac thieu `reason`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Thieu permission `catalog:publish`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Khong tim thay.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
...
  securitySchemes:
    webSession:
      type: apiKey
      in: header
      name: X-Session-Token
      description: |
        Web session token do BFF forward (từ cookie HttpOnly). Control Plane tra hash
        trong `web_sessions`, kiểm chưa hết hạn/chưa thu hồi, rồi resolve accountId.
        Token thô KHÔNG được log. Đây là kênh của NGƯỜI DÙNG — khác `serviceIdentity` (M2M).
    userSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token user do Auth0 phát, được BFF chuyển tiếp.
        **CHƯA dùng ở P1** — wiring Auth0 thật là P2 và phụ thuộc DEC-B03 (tenant thật).
    serviceIdentity:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        M2M token của backend ứng dụng (Data Plane). Mỗi backend có identity RIÊNG,
        audience cụ thể và exact resource scope. Không dùng chung credential.
        **CHƯA dùng ở P1.**
  schemas:
...
    PublicApplication:
      type: object
      required: [id, key, displayName, description, imageUrl, launchUrl]
      description: |
        View cho nguoi dung cuoi. CO Y khong co `status`: nguoi dung chi thay app `active`,
        nen truong do khong mang thong tin gi ngoai viec lo ra he thong co trang thai khac.
      properties:
        id: { type: string, format: uuid }
        key: { type: string }
        displayName: { type: string }
        description: { type: [string, 'null'] }
        imageUrl: { type: [string, 'null'] }
        launchUrl:
          type: string
          description: |
            Da qua chinh sach URL va da chuan hoa. Frontend dung NGUYEN chuoi nay, khong
            ghep host/path tu input nguoi dung.
    AdminApplication:
      type: object
      required: [id, key, displayName, description, imageUrl, launchUrl, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        key: { type: string }
        displayName: { type: string }
        description: { type: [string, 'null'] }
        imageUrl: { type: [string, 'null'] }
        launchUrl: { type: string }
        status: { type: string, enum: [draft, active, inactive] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ErrorEnvelope:
      type: object
      description: |
        Hình dạng lỗi ổn định cho MỌI lỗi rời khỏi API, kể cả lỗi không lường trước.

        Ràng buộc bảo mật: không bao giờ chứa stack trace, secret, connection string
        hay chi tiết nội bộ. Với lỗi 5xx, `message` luôn là thông điệp chung — chi tiết
        thật chỉ đi vào log phía server, truy được bằng `correlationId`.
      required: [code, message, correlationId]
      additionalProperties: false
      properties:
        code:
          type: string
          description: |
            Mã máy đọc được, ổn định, không phụ thuộc câu chữ giao diện.
            Client bắt lỗi theo trường này.
          pattern: '^[A-Z][A-Z0-9_]*$'
          examples:
            - INTERNAL_ERROR
            - VALIDATION_FAILED
            - SERVICE_UNAVAILABLE
        message:
          type: string
          description: Thông điệp an toàn để hiển thị.
          maxLength: 500
        correlationId:
          type: string
          format: uuid
          description: |
            Truy vết xuyên BFF -> API -> worker -> log.
            Nhận từ header `X-Correlation-Id` nếu hợp lệ, ngược lại server sinh mới (UUIDv7).
            **Không phải** idempotency key — đừng dùng lẫn.
        details:
          type: object
          description: Chi tiết bổ sung theo allowlist. Không bao giờ chứa dữ liệu nhạy cảm.
          additionalProperties: true
  parameters:
...
    ApplicationId:
      name: applicationId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
security: []